requirements.txt
wheel>=0.42.0
setuptools>=69.0.2
numpy==1.24.3
PySide6==6.6.1
Pillow>=10.1.0
pytest>=8.0.0
rembg>=2.0.50
opencv-python>=4.8.0